refactoring dist into an object
Includes other project refactoring to separate and isolate concerns.
