+++ conda/misc.py
@@ -8,18 +8,22 @@ import re
 import shutil
 import sys
 from collections import defaultdict
+from conda.base.constants import DEFAULTS
+from conda.core.package_cache import cached_url, is_fetched, is_extracted, find_new_location
+from conda.models.dist import Dist
 from os.path import (abspath, dirname, expanduser, exists,
                      isdir, isfile, islink, join, relpath, curdir)
 
-from .api import get_index
+from conda.core.index import get_index
 from .base.context import context
 from .common.url import path_to_url, is_url
 from .compat import iteritems, itervalues
 from .models.channel import Channel
 from .exceptions import (CondaFileNotFoundError, ParseError, MD5MismatchError,
                          PackageNotFoundError, CondaRuntimeError)
-from .install import (name_dist, linked as install_linked, is_fetched, is_extracted, is_linked,
-                      linked_data, find_new_location, cached_url, dist2filename)
+# from .install import (name_dist, linked as install_linked, is_fetched, is_extracted, is_linked,
+#                       linked_data, find_new_location, cached_url, dist2filename)
+from .core.linked_data import linked as install_linked, is_linked, linked_data
 from .instructions import RM_FETCHED, FETCH, RM_EXTRACTED, EXTRACT, UNLINK, LINK, SYMLINK_CONDA
 from .plan import execute_actions
 from .resolve import Resolve, MatchSpec
@@ -39,7 +43,7 @@ def conda_installed_files(prefix, exclude_self_build=False):
         meta = is_linked(prefix, dist)
         if exclude_self_build and 'file_hash' in meta:
             continue
-        res.update(set(meta['files']))
+        res.update(set(meta.get('files', ())))
     return res
 
 url_pat = re.compile(r'(?:(?P<url_p>.+)(?:[/\\]))?'
@@ -49,7 +53,7 @@ def explicit(specs, prefix, verbose=False, force_extract=True, index_args=None, 
     actions = defaultdict(list)
     actions['PREFIX'] = prefix
     actions['op_order'] = RM_FETCHED, FETCH, RM_EXTRACTED, EXTRACT, UNLINK, LINK, SYMLINK_CONDA
-    linked = {name_dist(dist): dist for dist in install_linked(prefix)}
+    linked = {dist.dist_name: dist for dist in install_linked(prefix)}
     index_args = index_args or {}
     index = index or {}
     verifies = []
@@ -78,17 +82,17 @@ def explicit(specs, prefix, verbose=False, force_extract=True, index_args=None, 
         is_cache = prefix is not None
         if is_cache:
             # Channel information from the cache
-            schannel = 'defaults' if prefix == '' else prefix[:-2]
+            schannel = DEFAULTS if prefix == '' else prefix[:-2]
         else:
             # Channel information from the URL
             channel, schannel = Channel(url).url_channel_wtf
-            prefix = '' if schannel == 'defaults' else schannel + '::'
+            prefix = '' if schannel == DEFAULTS else schannel + '::'
 
         fn = prefix + fn
-        dist = fn[:-8]
+        dist = Dist.from_string(fn[:-8])
         # Add explicit file to index so we'll be sure to see it later
         if is_local:
-            index[fn] = {'fn': dist2filename(fn), 'url': url, 'md5': md5}
+            index[dist] = {'fn': dist.to_filename(), 'url': url, 'md5': md5}
             verifies.append((fn, md5))
 
         pkg_path = is_fetched(dist)
@@ -109,18 +113,18 @@ def explicit(specs, prefix, verbose=False, force_extract=True, index_args=None, 
         if not dir_path:
             if not pkg_path:
                 pkg_path, conflict = find_new_location(dist)
-                pkg_path = join(pkg_path, dist2filename(dist))
+                pkg_path = join(pkg_path, dist.to_filename())
                 if conflict:
                     actions[RM_FETCHED].append(conflict)
                 if not is_local:
-                    if fn not in index or index[fn].get('not_fetched'):
+                    if dist not in index or index[dist].get('not_fetched'):
                         channels.add(schannel)
                     verifies.append((dist + '.tar.bz2', md5))
                 actions[FETCH].append(dist)
             actions[EXTRACT].append(dist)
 
         # unlink any installed package with that name
-        name = name_dist(dist)
+        name = dist.dist_name
         if name in linked:
             actions[UNLINK].append(linked[name])
 
@@ -175,7 +179,7 @@ def explicit(specs, prefix, verbose=False, force_extract=True, index_args=None, 
 
     # Finish the MD5 verification
     for fn, md5 in verifies:
-        info = index.get(fn)
+        info = index.get(Dist.from_string(fn))
         if info is None:
             raise PackageNotFoundError(fn, "no package '%s' in index" % fn)
         if md5 and 'md5' not in info:
@@ -320,13 +324,13 @@ def clone_env(prefix1, prefix2, verbose=True, quiet=False, index_args=None):
         index = get_index(**index_args)
         r = Resolve(index, sort=True)
         for dist in unknowns:
-            name = name_dist(dist)
-            fn = dist2filename(dist)
+            name = dist.dist_name
+            fn = dist.to_filename()
             fkeys = [d for d in r.index.keys() if r.index[d]['fn'] == fn]
             if fkeys:
                 del drecs[dist]
-                dist = sorted(fkeys, key=r.version_key, reverse=True)[0]
-                drecs[dist] = r.index[dist]
+                dist_str = sorted(fkeys, key=r.version_key, reverse=True)[0]
+                drecs[Dist.from_string(dist_str)] = r.index[dist_str]
             else:
                 notfound.append(fn)
     if notfound:
@@ -338,7 +342,7 @@ def clone_env(prefix1, prefix2, verbose=True, quiet=False, index_args=None):
     # Assemble the URL and channel list
     urls = {}
     for dist, info in iteritems(drecs):
-        fkey = dist + '.tar.bz2'
+        fkey = dist
         if fkey not in index:
             info['not_fetched'] = True
             index[fkey] = info
