+++ tests/test_install.py
@@ -15,9 +15,8 @@ from conda import install
 from conda.base.context import context
 from conda.common.disk import move_path_to_trash
 from conda.compat import text_type
-from conda.fetch import download
-from conda.install import (FileMode, PaddingError, binary_replace, dist2dirname, dist2filename,
-                           dist2name, dist2pair, dist2quad, name_dist, on_win,
+from conda.core.package_cache import download
+from conda.install import (FileMode, PaddingError, binary_replace, on_win,
                            read_no_link, update_prefix, warn_failed_remove, yield_lines,
                            _PaddingError)
 from contextlib import contextmanager
@@ -484,24 +483,24 @@ class remove_readonly_TestCase(unittest.TestCase):
 #         self.assertEqual(2, mocks['rmtree'].call_count)
 
 
-def test_dist2():
-    for name in ('python', 'python-hyphen', ''):
-        for version in ('2.7.0', '2.7.0rc1', ''):
-            for build in ('0', 'py27_0', 'py35_0+g34fe21', ''):
-                for channel in ('defaults', 'test', 'test-hyphen', 'http://bremen',
-                                'https://anaconda.org/mcg', '<unknown>'):
-                    dist_noprefix = name + '-' + version + '-' + build
-                    quad = (name, version, build, channel)
-                    dist = dist_noprefix if channel == 'defaults' else channel + '::' + dist_noprefix
-                    for suffix in ('', '.tar.bz2', '[debug]', '.tar.bz2[debug]'):
-                        test = dist + suffix
-                        assert dist2quad(test) == quad
-                        assert dist2pair(test) == (channel, dist_noprefix)
-                        assert dist2name(test) == name
-                        assert name_dist(test) == name
-                        assert dist2dirname(test) == dist_noprefix
-                        assert dist2filename(test) == dist_noprefix + '.tar.bz2'
-                        assert dist2filename(test, '') == dist_noprefix
+# def test_dist2():
+#     for name in ('python', 'python-hyphen', ''):
+#         for version in ('2.7.0', '2.7.0rc1', ''):
+#             for build in ('0', 'py27_0', 'py35_0+g34fe21', ''):
+#                 for channel in ('defaults', 'test', 'test-hyphen', 'http://bremen',
+#                                 'https://anaconda.org/mcg', '<unknown>'):
+#                     dist_noprefix = name + '-' + version + '-' + build
+#                     quad = (name, version, build, channel)
+#                     dist = dist_noprefix if channel == 'defaults' else channel + '::' + dist_noprefix
+#                     for suffix in ('', '.tar.bz2', '[debug]', '.tar.bz2[debug]'):
+#                         test = dist + suffix
+#                         assert dist2quad(test) == quad
+#                         assert dist2pair(test) == (channel, dist_noprefix)
+#                         assert dist2name(test) == name
+#                         assert name_dist(test) == name
+#                         assert dist2dirname(test) == dist_noprefix
+#                         assert dist2filename(test) == dist_noprefix + '.tar.bz2'
+#                         assert dist2filename(test, '') == dist_noprefix
 
 
 def _make_lines_file(path):
